fix: Convert to sendLater() function instead of custom event system.
